Move single instance check logic into SingleInstanceChecker.py
